+++ test/td_bg/business_concepts/search/encode_test.exs
@@ -282,16 +282,16 @@ defmodule TdBg.BusinessConcepts.Search.EncodeTest do
                "text_area" => "default_foo",
                "text_area_es" => "bar_translatable",
                "text_input_es" => "foo_translatable",
-               "enriched_text" => "",
-               "enriched_text_es" => "",
                "Hierarchie2" => "",
                "User Group" => "",
                "basic_switch" => "",
                "default_dependency" => "1.1",
-               "empty test" => "",
                "multiple_values" => [""],
                "text_input" => "",
-               "user1" => ""
+               "user1" => "",
+               "empty test" => "",
+               "enriched_text" => "",
+               "enriched_text_es" => ""
              } == content
     end
 
@@ -318,8 +318,12 @@ defmodule TdBg.BusinessConcepts.Search.EncodeTest do
           "origin" => "user"
         },
         "basic_list" => %{"value" => "1", "origin" => "user"},
-        "Identificador" => %{"value" => "foo", "origin" => "user"},
-        "text_area" => %{"value" => "default_foo", "origin" => "user"}
+        "text_input" => %{"value" => "original_foo", "origin" => "user"}
+      }
+
+      i18n_content = %{
+        "text_input" => %{"value" => "foo_translatable", "origin" => "user"},
+        "text_area" => %{"value" => nil, "origin" => "user"}
       }
 
       %{id: bcv_id} =
@@ -332,11 +336,6 @@ defmodule TdBg.BusinessConcepts.Search.EncodeTest do
         )
         |> Repo.preload(business_concept: :shared_to)
 
-      i18n_content = %{
-        "text_input" => %{"value" => "foo_translatable", "origin" => "user"},
-        "text_area" => %{"value" => "bar_translatable", "origin" => "user"}
-      }
-
       i18n_name = "i18n_name"
 
       insert(:i18n_content,
@@ -346,27 +345,101 @@ defmodule TdBg.BusinessConcepts.Search.EncodeTest do
         name: i18n_name
       )
 
-      assert %{name_es: ^i18n_name, content: content} = Document.encode(bcv)
+      assert %{name_es: ^i18n_name, content: content} =
+               Document.encode(bcv)
 
       assert %{
-               "Identificador" => "foo",
                "basic_list" => "1",
+               "text_input" => "original_foo",
+               "text_input_es" => "foo_translatable",
+               "text_area" => "",
+               "text_area_es" => "",
                "df_description" => "enrich text",
                "df_description_es" => "enrich text",
-               "text_area" => "default_foo",
-               "text_area_es" => "bar_translatable",
-               "text_input_es" => "foo_translatable",
-               "enriched_text" => "",
-               "enriched_text_es" => "",
+               "empty test" => "",
                "Hierarchie2" => "",
                "User Group" => "",
                "basic_switch" => "",
+               "enriched_text" => "",
+               "enriched_text_es" => "",
                "default_dependency" => "1.1",
-               "empty test" => "",
                "multiple_values" => [""],
-               "text_input" => "",
                "user1" => ""
              } == content
     end
+  end
+
+  test "when encodes empty translatable fields, assigns the default value" do
+    busines_concept_content = %{
+      "df_description" => %{
+        "value" => %{
+          "document" => %{
+            "nodes" => [
+              %{
+                "nodes" => [
+                  %{
+                    "marks" => [],
+                    "object" => "text",
+                    "text" => "enrich text"
+                  }
+                ],
+                "object" => "block",
+                "type" => "paragraph"
+              }
+            ]
+          }
+        },
+        "origin" => "user"
+      },
+      "basic_list" => %{"value" => "1", "origin" => "user"},
+      "text_input" => %{"value" => "original_foo", "origin" => "user"},
+      "text_area" => %{"value" => "original_bar", "origin" => "user"}
+    }
+
+    i18n_content = %{
+      "text_input" => %{"value" => "", "origin" => "user"},
+      "text_area" => %{"value" => nil, "origin" => "user"}
+    }
+
+    %{id: bcv_id} =
+      bcv =
+      :business_concept_version
+      |> insert(
+        content: busines_concept_content,
+        name: "Concept Name",
+        type: @template_name
+      )
+      |> Repo.preload(business_concept: :shared_to)
+
+    i18n_name = "i18n_name"
+
+    insert(:i18n_content,
+      business_concept_version_id: bcv_id,
+      content: i18n_content,
+      lang: "es",
+      name: i18n_name
+    )
+
+    assert %{name_es: ^i18n_name, content: content} =
+             Document.encode(bcv)
+
+    assert %{
+             "basic_list" => "1",
+             "empty test" => "",
+             "text_input" => "original_foo",
+             "text_input_es" => "original_foo",
+             "text_area" => "original_bar",
+             "text_area_es" => "original_bar",
+             "df_description" => "enrich text",
+             "df_description_es" => "enrich text",
+             "Hierarchie2" => "",
+             "User Group" => "",
+             "basic_switch" => "",
+             "enriched_text" => "",
+             "enriched_text_es" => "",
+             "default_dependency" => "1.1",
+             "multiple_values" => [""],
+             "user1" => ""
+           } == content
   end
 end